Moving argparse require out of pip-requires
argparse is included in 2.7+. We should only be installing it
if we are running Python 2.6.

Side note, the argparse version in pip needs to be repackaged
to conform to pip 1.5+ standards. Otherwise, we need to also
look into adding workarounds for that as well.

:: edit ::
Changed my mind on where I wanted to put this.

Change-Id: Idfa8ca93b105aebbc3aecbddcca7caee97192912