Add JIRA skip_on_issue support.
A new 'JIRA' section should be added to your *.config file.
This section only requires a field 'server' containing the address
of a JIRA server. A JIRA issue has a unique id among all projects,
'<project-id>-<issue-number>', so we don't need a project field in
the config file.

Issues are skipped based on the value of the issue's 'status' field.
JIRA allows you to define custom statuses, so the config file
optionally accepts a 'closed_statuses' field, which should be a JSON
formatted list of strings (to work with json.loads). A JIRA issue is
considered closed only if its status is in the 'closed_statuses' list.

Example config section:
    [JIRA]
    server=http://my.jira.server
    closed_statuses=["Closed", "Resolved", "Merged"]

Change-Id: I974a061b16c33089cb2c31fd0ed16b90a9b4240a