**Execute shell command fix**
The execute_shell_command opens a chanel, gets a pseudo terminal and then runs command. Enables running sudo commands.
The current implementation waits for a prompt and then returns the output. If no prompt is given, it returns whatever output it could fetch in the first attempt to recv() which could be different everytime.

My fix:
Changes the way shell command is executed to use chan.exec_command() instead of "send" to run the command. This way we can get the exit status of the command just run on the chanel and reliably terminate the client.
Also I'm verifying if the chanel is actually closed which indicates that the command execution is complete as per the docs.

Change-Id: I82bc2728c70f21d8837b0ff676a31d16f077ae5c