Update skip_on_issue to support JIRA labels
In the config file, your JIRA section can look like:
    [JIRA]
    server=http://my.jira.server
    closed_statuses=["Closed", "Resolved"]
    closed_labels=["Test", "Prod"]
This affects any test cases decorated with skip_on_issue('jira', ...)

* If closed_statuses is in your JIRA section, the test case is skipped
unless the JIRA issue's status is in the closed_statuses list
* If closed_labels is in your JIRA section, the test is skipped unless
any of the issue's labels are in the closed_labels list
* There are no longer any default values. If an empty closed_* list is
provided, the test will always skip. If a closed_* list is absent then
the corresponding status or label check is ignored.

Change-Id: Ifb8dd397b06a1fe59fcb113aa0570451f5f83237