Update skip_on_issue to support Launchpad tags
The Launchpad section of your config can look like:
    [Launchpad]
    project=barbican
    use_status=False
    closed_tags=["test", "prod"]
This affects functions decorated with skip_on_issue('launchpad', ...)

* Set use_status to False to ignore the status of the bug. This
defaults to True if absent for backwards compatibility.
* A bug is open if none of its tags is in closed_tags. If closed_tags
is absent, then the bug's tags are ignored.

Additionally, a get_json method is added to all data sources in the
data_interfaces module.

Change-Id: I7a76220a4b7da62865b1c2f7cd08414e9dbabfae