Removed TestEnvManager
Removed TestEnvManager because it was causing circular dependencies.
Orignal code had ConfigSectionInterface relying on TestEnvManger via
environmental variables relying on EngineConfig config values which
relied on ConfigSectionInterface.  This made it mandatory to init
the TestEnvManager prior to doing any thing with configs.  Now
the dependency chain is
ConfigSectioninterface -> EngineConfig -> Bare minimum config parser
-> DirectoryManager
The downsides of this design are, anything that is in the managers.py
module cannot use the engine config without creating a circular
dependency so is a future pull request it would be good to
separate the features required to create the engine config into
a different module from the rest of the managers.

Change-Id: I783555348220c335b0d9b3f38c5fdc4d2a7cf0cb