Add individual test times to xml report.
Currently, the xml report only reports the total test execution time, and do not provide data on the time
taken for individual test cases. This change add a _duration attribute to the base test case that is then
populated during tearDown and added to a 'test_time' attribute on the result object. The duration is
determined using a the same mechanism as the test_metrics reporting. The time is reported in seconds.

Change-Id: I41931e12e13e073debd37403fcc0030c26041fb0