Change to DataDrivenClass generator to fail when DSL does not generate data.
If a DataSetList is passed into a test that does not generate or
contain data the current behavior is that the test suites passes.
This proposal is to change that behavior. If a DSL cannot generate
data the test suite as a whole should still run but it should
fail. Otherwise teams using tooling to track and manage test results
will have false positives with mysteriously missing tests.

Added Metatests to check basic functionality.

Change-Id: I3430e9573c3e99b80f25e7577709144d49c8eae2