Added Safety and clarity around tests setups missing Super()
Test setUp methods that do not have a 'super()' call will
fail and cause the entire suite to fail despite tests
passing. Runner failures should, whenever possible, not
interfere with test results.

This is to add a try/except around getting the '_duration'
value for tests. If a test uses a setUp() method without super
it will recieve 'NaN' as a 'duration' and the user will
recieve a warning in console output but will not recieve
failures.

This is considered a safety mechanism intended to provide
test writers time to update their tests using 'setUp()' to
correctly call super().

Additionally added some newlines around the warning so that it is
very clear that there is an issue.

Change-Id: I32d2352ff7081b84383835bcc8812e37735c68d2