WIP - Added mechanism to allow retries around Requests Lib exceptions
Often exceptions caused by the Request Library are either:
* Impacting, in which case they are consistent
* Non-Impacting, in which case they are sporadic and cause false
  failures

In tests that do API testing using the HTTP plugin, there is a use
case for retrying Request Library exceptions as some of these
failures can be caused by issues unrelated to tests and the system
under test. This PR provides a mechanism to configure how many of
these errors a test runner is willing to ignore.

Change-Id: I2d15095ad435b151df56f0a23f068f0fae662c37